Authentication-KeyCloak/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
requests
python-dotenv
PyJWT